Move to `type_check_controller.rb`
