+++ sig/steep/server/custom_methods.rbs
@@ -49,7 +49,7 @@ module Steep
       module TypeCheck__Start
         METHOD: String
 
-        type params = Master::TypeCheckRequest::json
+        type params = TypeCheckRequest::json
 
         def self.notification: (params) -> untyped
       end
