+++ test/master_test.rb
@@ -8,8 +8,8 @@ class MasterTest < Minitest::Test
   include Steep
 
   Master = Server::Master
-  TypeCheckController = Master::TypeCheckController
-  TypeCheckRequest = Master::TypeCheckRequest
+  TypeCheckController = Server::TypeCheckController
+  TypeCheckRequest = Server::TypeCheckRequest
   WorkDoneProgress = Server::WorkDoneProgress
 
   DEFAULT_CLI_LSP_INITIALIZE_PARAMS = Drivers::Utils::DriverHelper::DEFAULT_CLI_LSP_INITIALIZE_PARAMS
@@ -50,7 +50,7 @@ end
       progress = master.work_done_progress("guid")
       master.start_type_check(last_request: nil, progress: progress, report_progress_threshold: 0, needs_response: true)
 
-      assert_instance_of Server::Master::TypeCheckRequest, master.current_type_check_request
+      assert_instance_of Server::TypeCheckRequest, master.current_type_check_request
 
       jobs = flush_queue(master.write_queue)
 
@@ -116,7 +116,7 @@ end
       progress = master.work_done_progress("guid")
       master.start_type_check(last_request: nil, progress: progress, report_progress_threshold: 0, needs_response: true)
 
-      assert_instance_of Server::Master::TypeCheckRequest, master.current_type_check_request
+      assert_instance_of Server::TypeCheckRequest, master.current_type_check_request
 
       jobs = flush_queue(master.write_queue)
 
@@ -301,7 +301,7 @@ end
       progress = master.work_done_progress("guid")
       master.start_type_check(last_request: nil, progress: progress, report_progress_threshold: 0, needs_response: true)
 
-      assert_instance_of Server::Master::TypeCheckRequest, master.current_type_check_request
+      assert_instance_of Server::TypeCheckRequest, master.current_type_check_request
 
       flush_queue(master.write_queue)
